# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "contact": {}
    },
    "paths": {
        "/artist-info": {
            "get": {
                "description": "Get top tracks based on specified parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GetArtistInfo"
                ],
                "summary": "Show Artist info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The method parameter",
                        "name": "method",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The country parameter",
                        "name": "artist",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The API key parameter",
                        "name": "api_key",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The format parameter",
                        "name": "format",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "Return"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/image-info": {
            "get": {
                "description": "Get top tracks based on specified parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GetImageInfo"
                ],
                "summary": "Show image info",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The method parameter",
                        "name": "method",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The country parameter",
                        "name": "artist",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The API key parameter",
                        "name": "api_key",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The format parameter",
                        "name": "format",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "Return"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        },
        "/top-tracks": {
            "get": {
                "description": "Get top tracks based on specified parameters",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "GetTopTracks"
                ],
                "summary": "Show top tracks",
                "parameters": [
                    {
                        "type": "string",
                        "description": "The method parameter",
                        "name": "method",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The country parameter",
                        "name": "country",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The API key parameter",
                        "name": "api_key",
                        "in": "query",
                        "required": true
                    },
                    {
                        "type": "string",
                        "description": "The format parameter",
                        "name": "format",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "Return"
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not Found"
                    },
                    "500": {
                        "description": "Internal Server Error"
                    }
                }
            }
        }
    }
}